# More performance optimizations