cc @cw9 @jeromefroe @prateek @dgromov 

This PR adds more performance optimizations for the agg tier, including:
* Make elem lock per time bucket. Previously we used to lock the entire element during writing or flushing, regardless of which time bucket we are writing too, resulting in higher lock contention than necessary. This PR now makes the lock per time bucket and thus more granular, so that writes to the current time bucket don't contend with reads from the previous time bucket during flushing.
* Add option to support periodic insertion and compression. For high frequency timers in large timer batches, it is very important to add timers as fast as possible. By only performing insertion and compression, I noticed a p99 write latency drop from 400ms to 30ms for higher frequency timer writes.
* Add option to support splitting large timer batches into smaller ones. This provides the flexibility to write a large batch of timers in smaller batches, therefore making each batched write hold locks for shorter period of time and thus reduce lock contention with flushes. I noticed significant flush time reduction for high frequency timers in large batches due to batch splitting (8s -> 1s).